patitas/requirements.txt
django
djangorestframework
cloudinary
django-cors-headers
python-decouple
djangorestframework-simplejwt
psycopg2
gunicorn